requirements.txt
asgiref==3.3.4
bleach==3.3.0
Django==3.2
django-classy-tags==2.0.0
django-filter==2.4.0
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-mptt==0.11.0
django-nyt==1.2
django-sekizai==2.0.0
djangorestframework==3.12.4
Markdown==3.2.2
packaging==20.9
Pillow==8.2.0
pyparsing==2.4.7
pytz==2021.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
webencodings==0.5.1
wiki==0.7.5
